backend/requirement.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
frozendict==2.4.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
lxml==5.3.1
multitasking==0.0.11
nsepy==0.8
numpy==2.2.3
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
yfinance==0.2.53
